Add files via upload
In this project I utilize real world voter data in which voters are asked to rank 5 presidential candidates for the 2016 election to predict the candidate with the highest probability of winning. I use various methods such as plurality count, Borda count, average scores, and the PageRank algorithm (after creating a network representation of the data).