Add files via upload
In this algorithm I use mathematical methods from linear algebra such as QR factorization and SVD decomposition of matrices (images) to create an efficient and accurate handwritten digit classifying algorithm. 

I do not own the data used in this project, it was provided to me by the University of California, Davis for educational purposes.